autojob/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

apify-client==1.6.2
cupy==13.0.0
numpy==1.26.3
python-dateutil==2.8.2
requests==2.31.0
skillNer==1.0.3
spacy==3.7.2
textacy==0.13.0
tqdm==4.66.1
transformers==4.37.0
